# quickstart/common/openapi.yaml
openapi: 3.0.0
info:
  title: Backend API
  description: API to manage AppInstalls and Licenses
  version: 1.0.0
paths:
...
  /license-renewal-requests:
    get:
      summary: List all LicenseRenewalRequests
      operationId: listLicenseRenewalRequests
      responses:
        '200':
          description: A list of LicenseRenewalRequests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LicenseRenewalRequest'
        '400':
          description: Invalid input
  /license-renewal-requests/{contractId}/complete-renewal:
    post:
      summary: Complete the License Renewal
      operationId: completeLicenseRenewal
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: string
          description: The contract ID of the LicenseRenewalRequest
        - name: commandId
          in: query
          required: true
          schema:
            type: string
          description: The command ID for the ledger transaction
      responses:
        '200':
          description: License renewed successfully
        '400':
          description: Invalid input
...
components:
  schemas:
...
    LicenseRenewalRequest:
      type: object
      properties:
        contractId:
          type: string
        provider:
          type: string
        user:
          type: string
        dso:
          type: string
        licenseNum:
          type: integer
        licenseFeeCc:
          type: number
        licenseExtensionDuration:
          type: string
          description: "RelTime representing how long the license should be extended."
        reference:
          type: string
...